[refactor] Extract parent class from Question and Answer
