+++ src/main/java/nextstep/qna/domain/Question.java
@@ -6,22 +6,9 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Question {
-    private Long id;
-
+public class Question extends Content {
     private String title;
-
-    private String contents;
-
-    private NsUser writer;
-
     private List<Answer> answers = new ArrayList<>();
-
-    private boolean deleted = false;
-
-    private LocalDateTime createdDate = LocalDateTime.now();
-
-    private LocalDateTime updatedDate;
 
     public Question() {
     }
@@ -31,15 +18,10 @@ public class Question {
     }
 
     public Question(Long id, NsUser writer, String title, String contents) {
-        this.id = id;
-        this.writer = writer;
+        super(id, writer, contents);
         this.title = title;
-        this.contents = contents;
     }
 
-    public Long getId() {
-        return id;
-    }
 
     public String getTitle() {
         return title;
@@ -50,35 +32,14 @@ public class Question {
         return this;
     }
 
-    public String getContents() {
-        return contents;
-    }
-
     public Question setContents(String contents) {
         this.contents = contents;
         return this;
     }
 
-    public NsUser getWriter() {
-        return writer;
-    }
-
     public void addAnswer(Answer answer) {
         answer.toQuestion(this);
         answers.add(answer);
-    }
-
-    public boolean isOwner(NsUser loginUser) {
-        return writer.equals(loginUser);
-    }
-
-    public Question setDeleted(boolean deleted) {
-        this.deleted = deleted;
-        return this;
-    }
-
-    public boolean isDeleted() {
-        return deleted;
     }
 
     public List<Answer> getAnswers() {
